// src/types.ts
  export type Items = Item[];

  export interface Item {
    id: number;
    game_id: number;
    name: string;
    display_name: string;
    image: string;
    price?: number;
    total_price: number;
    slot_type?: string;
    rarity?: string;
    aggression_type?: string;
    hero_class?: string;
    required_level?: number;
    effects: ItemEffect[];
    requirements: string[];
    build_paths: string[];
    stats?: ItemStats;
  }

  export interface ItemEffect {
    name: string;
    active: boolean;
    game_description: string;
    menu_description: string;
  }

  export interface ItemStats {
    magical_power?: number;
    physical_power?: number;
    critical_chance?: number;
    attack_speed?: number;
    omnivamp?: number;
    ability_haste?: number;
    max_health?: number;
    tenacity?: number;
    magical_armor?: number;
    magical_lifesteal?: number;
    max_mana?: number;
    mana_regeneration?: number;
    physical_penetration?: number;
    lifesteal?: number;
    physical_armor?: number;
    movement_speed?: number;
    magical_penetration?: number;
    health_regeneration?: number;
    heal_shield_increase?: number;
    gold_per_second?: number;
  }